pyproject.toml
[project]
name = "trading-research"
version = "0.1.0"
description = "Crypto trading ML research pipeline"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=4.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
]

[project.scripts]
run-experiment = "run_experiment:main"
